+++ src/main/java/nextstep/security/context/SecurityContextRepository.java
@@ -2,22 +2,9 @@ package nextstep.security.context;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
 
-public class SecurityContextRepository {
-    public static final String SPRING_SECURITY_CONTEXT_KEY = "SPRING_SECURITY_CONTEXT";
+public interface SecurityContextRepository {
+    SecurityContext loadContext(HttpServletRequest request);
 
-    public SecurityContext loadContext(HttpServletRequest request) {
-        HttpSession session = request.getSession(false);
-        if (session == null) {
-            return null;
-        }
-
-        return (SecurityContext) session.getAttribute(SPRING_SECURITY_CONTEXT_KEY);
-    }
-
-    public void saveContext(SecurityContext context, HttpServletRequest request, HttpServletResponse response) {
-        HttpSession session = request.getSession();
-        session.setAttribute(SPRING_SECURITY_CONTEXT_KEY, context);
-    }
+    void saveContext(SecurityContext context, HttpServletRequest request, HttpServletResponse response);
 }
